Move Mocks in right folders
